+++ QLDA.Domain/DTOs/DashboardGiaiNganDto.cs
@@ -18,9 +18,9 @@ public class DashboardGiaiNganTheoNguonVonDto {
 public class DashboardChiTietGiaiNganDto {
     public string? TenDuAn { get; set; }
     /// <summary>Số tiền đã giải ngân (ThanhToan.GiaTri)</summary>
-    public long? GiaTriGiaiNgan { get; set; }
+    public decimal? GiaTriGiaiNgan { get; set; }
     /// <summary>Giá trị hợp đồng (HopDong.GiaTri)</summary>
-    public long? GiaTriHopDong { get; set; }
+    public decimal? GiaTriHopDong { get; set; }
     public DateTimeOffset? Ngay { get; set; }
     public string? TrangThaiGiaiNgan { get; set; }
 }
